auth-server: api-handlers: Do not re-route whitelisted api keys (#279)
An external match request will be routed to the global matching pool if
the bot server's daily execution cost rate limit has been exceeded on
the requested pair.

The exception introduced in this PR is if the requesting user holds a
whitelisted key, in which case their orders are never routed to the
global matching pool.