update price reporter telemetry config (#438)
In this PR, we transition the price reporter to using `clap` with `env` attributes to fit to existing patterns for configuring our extensions. This makes for a cleaner invocation of the now-standardized `configure_telemetry_with_metrics_config` method exported from the relayer repo.

In the next PR, we'll add some additional telemetry to the price reporter to be used for debugging some websocket reliability issues. Testing will be done subsequent to that PR.