price-reporter: price_stream_manager: treat stream termination as error (#445)
In this PR, we handle exchange connection price stream closures correctly by erroring out of `manage_connection` so that the outer `price_stream_task` falls through to the requisite reconnection logic.

This fixes a bug wherein closed websocket connections would only be reconnected to on subsequent errors in reading/writing messages (primarily, this would be due to keepalive messages that are sent at most 15s after the closure).

### Testing
- [x] Ran a local price reporter connected to a dummy websocket server where I triggered websocket connection closures. Asserted that before this change, exchange connections would remain idle & would reconnect only after the keepalive is attempted to be sent; and that after this change, reconnection was attempted as soon as the closure was detected.