price-reporter: exchange: coinbase: refresh ws subscription instead of fetching book (#446)
In this PR, we replace the refreshing of the Coinbase orderbook via an HTTP fetch with a resubscription to the `level2` channel for the given product. This results in a fresh snapshot being sent when the subscription is re-established, and ensures that no updates are dropped due to the refresh. Additionally, this removes the need for the `rehydration_lock` on the orderbook.

### Testing

- [x] Ran local price reporter, asserted that fresh snapshots are received due to resubscription and that price streams to clients experience no significant delays
- [x] Deployed to testnet, where we see intra-cluster price deviations decrease considerably