darkpool-indexer: db: schema: add account ID cols & expected nullifiers table (#455)
In this PR, we update the schema to include an account ID column for all relevant tables, and to include an "expected nullifiers" table. The former is used to enable higher-level groupings of user state than on a per-EOA basis; the latter is used to avoid having empty, uninitialized state in the generic state objects table when precomputing new object nullifiers to look out for.

I modified the original migration instead of creating a new one since the migration hasn't been applied to any deployed infrastructure.